refactor(scrape): files structure
