+++ scrape/num-of-items-in-data.js
@@ -1,4 +1,4 @@
-const data = require('./data.json');
+const data = require('./data/data.json');
 
 for (let category in data) {
   console.log(`${category}: ${data[category].length} items`); // eslint-disable-line no-console
